all time stats for batting and pitching

+++ pcbl-royals/src/main/java/com/main/pcblroyals/bean/PitchingStatPlayerBean.java
@@ -95,7 +95,7 @@ public class PitchingStatPlayerBean extends PitchingStatBean implements Serializ
                             int losses,
                             int ties,
                             int saves,
-                            float inningsPitchedRaw,
+                            double inningsPitchedRaw,
                             int earnedRuns,
                             int totalRuns,
                             int strikeouts,
